refactor(models): rename architectures/ → multimodal/ + clean up stale moe comments
The `architectures/` directory under `ferrum-models/` exclusively held
multi-modal model implementations (BERT, CLIP, Whisper, Qwen3-TTS) —
NOT general "architectures" (those live in `models/llama_family.rs` /
`models/qwen3_moe.rs`). The old name suggested a flat dump of all
model architectures and obscured which directory owns the LLM hot
path. Rename to the more descriptive `multimodal/`.

Also cleans up two stale "Phase 2 deferred" comments in `moe/`:
- `moe/layer.rs` claimed `Qwen3MoeLayer` was the production MoE entry.
  Production goes through `models::qwen3_moe::Qwen3MoeLayerState` +
  the generic `crate::moe::moe_forward<B>` (already shipped). The
  `Qwen3MoeLayer<CpuBackend>` wrapper is now test-only.
- `moe/dispatch.rs::ExpertStack` claimed generic `moe_forward<B>` was
  deferred. It's been the production path since Phase 3 (line ~960).

No behaviour change. Mac local: `cargo check --workspace --all-targets
--features metal` ✅, `cargo test --workspace --features metal --lib`
13 suites / 330+ tests ✅ 0 failed.
