+++ crates/ferrum-models/src/executor/clip_executor.rs
@@ -20,7 +20,7 @@ use ferrum_types::{DataType, Device, FerrumError, ModelInfo, ModelType, Result};
 use tracing::info;
 
 use super::common;
-use crate::architectures::clip::ClipModelWrapper;
+use crate::multimodal::clip::ClipModelWrapper;
 use crate::image_processor::ClipImageProcessor;
 use crate::tensor_wrapper::CandleTensorWrapper;
