+++ crates/ferrum-models/src/multimodal/clip.rs
