+++ crates/ferrum-models/src/multimodal/mel_filters128.bin
