+++ crates/ferrum-models/src/multimodal/mel_filters80.bin
