+++ crates/ferrum-models/src/multimodal/mod.rs
