+++ crates/ferrum-models/src/multimodal/qwen3_tts_backend.rs
@@ -23,7 +23,7 @@ use ferrum_quantization::PrefixedLoader;
 use ferrum_types::Result;
 use std::collections::HashMap;
 
-use crate::architectures::qwen3_tts::TalkerConfig;
+use crate::multimodal::qwen3_tts::TalkerConfig;
 use crate::models::llama_family::{LlamaFamilyConfig, LlamaFamilyModel};
 
 /// Qwen3-TTS Talker, Model-as-Code implementation over `Backend<B>`.
