+++ crates/ferrum-models/src/multimodal/qwen3_tts_vocoder.rs
