+++ crates/ferrum-models/src/multimodal/speaker_encoder.rs
