+++ crates/ferrum-models/src/multimodal/speech_tokenizer_encoder.rs
