+++ crates/ferrum-models/src/multimodal/whisper.rs
